+++ srcs/raycasting.c
@@ -6,7 +6,7 @@
 /*   By: lguiller <lguiller@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/12 14:06:10 by lguiller          #+#    #+#             */
-/*   Updated: 2018/05/17 15:34:09 by lguiller         ###   ########.fr       */
+/*   Updated: 2018/05/21 11:23:10 by lguiller         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,9 +38,9 @@ static void	ft_fp_hori(t_ray *ray, t_player *p, char map[MAPY][MAPX], double a)
 	if (a == WEST || a == EAST || a == EAST2)
 		ray->fy = p->y;
 	else if (sin(a) >= 0)
-		ray->fy = (double)(ft_roundminf((int)p->y, (int)BLOCK_SIZE)) - LITTLE;
+		ray->fy = (double)(ft_roundminf(p->y, (int)BLOCK_SIZE)) - LITTLE;
 	else
-		ray->fy = (double)(ft_roundmsup((int)p->y, (int)BLOCK_SIZE));
+		ray->fy = (double)(ft_roundmsup(p->y, (int)BLOCK_SIZE));
 	ray->fx = p->x + (p->y - ray->fy) / tan(a);
 	ray->ya = (sin(a) >= 0) ? -BLOCK_SIZE : BLOCK_SIZE;
 	ray->xa = -ray->ya / tan(a);
@@ -53,9 +53,9 @@ static void	ft_fp_vert(t_ray *ray, t_player *p, char map[MAPY][MAPX], double a)
 	if (a == NORTH || a == SOUTH)
 		ray->fx = p->x;
 	else if (cos(a) < 0)
-		ray->fx = (double)(ft_roundminf((int)p->x, (int)BLOCK_SIZE)) - LITTLE;
+		ray->fx = ft_roundminf(p->x, (int)BLOCK_SIZE) - LITTLE;
 	else
-		ray->fx = (double)(ft_roundmsup((int)p->x, (int)BLOCK_SIZE));
+		ray->fx = ft_roundmsup(p->x, (int)BLOCK_SIZE);
 	ray->fy = p->y + (p->x - ray->fx) * tan(a);
 	ray->xa = (cos(a) < 0) ? -BLOCK_SIZE : BLOCK_SIZE;
 	ray->ya = -ray->xa * tan(a);
